+++ src/dialogs/welcomedialog.cpp
@@ -11,6 +11,17 @@ WelcomeDialog::WelcomeDialog(QWidget *parent) :
         MasterDialog(parent),
     ui(new Ui::WelcomeDialog) {
     ui->setupUi(this);
+
+    // replace ownCloud text
+    ui->subHeadlineLabel->setText(Utils::Misc::replaceOwnCloudText(
+            ui->subHeadlineLabel->text()));
+    ui->groupBox_2->setTitle(Utils::Misc::replaceOwnCloudText(
+            ui->groupBox_2->title()));
+    ui->label->setText(Utils::Misc::replaceOwnCloudText(ui->label->text()));
+    ui->label_4->setText(Utils::Misc::replaceOwnCloudText(ui->label_4->text()));
+    ui->ownCloudSettingsButton->setText(Utils::Misc::replaceOwnCloudText(
+            ui->ownCloudSettingsButton->text()));
+
     ui->finishButton->setEnabled(false);
     ui->backButton->setEnabled(false);
     ui->errorMessageLabel->setVisible(false);